+++ packages/contracts/contracts/core/DAO.sol
@@ -13,12 +13,15 @@ import "@openzeppelin/contracts/interfaces/IERC1271.sol";
 import "./component/CallbackHandler.sol";
 import "./permission/PermissionManager.sol";
 import "./IDAO.sol";
+import "./IEIP4824.sol";
+import {hasBit, flipBit} from "../utils/BitMap.sol";
 
 /// @title DAO
 /// @author Aragon Association - 2021
 /// @notice This contract is the entry point to the Aragon DAO framework and provides our users a simple and easy to use public interface.
 /// @dev Public API of the Aragon DAO framework.
 contract DAO is
+    IEIP4824,
     Initializable,
     IERC1271,
     ERC165StorageUpgradeable,
@@ -54,14 +57,24 @@ contract DAO is
     bytes32 public constant REGISTER_STANDARD_CALLBACK_PERMISSION_ID =
         keccak256("REGISTER_STANDARD_CALLBACK_PERMISSION");
 
+    /// @notice Only allows 256 actions to execute per tx.
+    uint256 internal constant MAX_ACTIONS = 256;
+
     /// @notice The [ERC-1271](https://eips.ethereum.org/EIPS/eip-1271) signature validator contract.
     IERC1271 public signatureValidator;
 
     /// @notice The address of the trusted forwarder verifying meta transactions.
     address private trustedForwarder;
 
+    /// @notice The [EIP-4824](https://eips.ethereum.org/EIPS/eip-4824) DAO uri.
+    string private _daoURI;
+
+    /// @notice Thrown if action length is more than MAX_ACTIONS.
+    error TooManyActions();
+
     /// @notice Thrown if action execution has failed.
-    error ActionFailed();
+    /// @param index Index of action in the array that failed.
+    error ActionFailed(uint256 index);
 
     /// @notice Thrown if the deposit or withdraw amount is zero.
     error ZeroAmount();
@@ -73,6 +86,10 @@ contract DAO is
 
     /// @notice Thrown if a native token withdraw fails.
     error NativeTokenWithdrawFailed();
+
+    /// @notice Emitted when a new DAO uri is set.
+    /// @param daoURI The new uri.
+    event NewURI(string daoURI);
 
     /// @dev Used to disallow initializing implementation contract by attacker for extra safety.
     constructor() {
@@ -90,13 +107,16 @@ contract DAO is
     function initialize(
         bytes calldata _metadata,
         address _initialOwner,
-        address _trustedForwarder
+        address _trustedForwarder,
+        string calldata daoURI_
     ) external initializer {
         _registerInterface(type(IDAO).interfaceId);
         _registerInterface(type(IERC1271).interfaceId);
+        _registerInterface(type(IEIP4824).interfaceId);
 
         _setMetadata(_metadata);
         _setTrustedForwarder(_trustedForwarder);
+        _setDaoURI(daoURI_);
         __PermissionManager_init(_initialOwner);
     }
 
@@ -160,21 +180,39 @@ contract DAO is
     }
 
     /// @inheritdoc IDAO
-    function execute(uint256 callId, Action[] calldata _actions)
+    function execute(
+        bytes32 callId,
+        Action[] calldata _actions,
+        uint256 allowFailureMap
+    )
         external
         override
         auth(address(this), EXECUTE_PERMISSION_ID)
-        returns (bytes[] memory)
+        returns (bytes[] memory execResults, uint256 failureMap)
     {
-        bytes[] memory execResults = new bytes[](_actions.length);
+        if (_actions.length > MAX_ACTIONS) {
+            revert TooManyActions();
+        }
 
-        for (uint256 i; i < _actions.length;) {
-            (bool success, bytes memory response) = _actions[i].to.call{value: _actions[i].value}(
+        execResults = new bytes[](_actions.length);
+
+        for (uint256 i = 0; i < _actions.length; ) {
+            address to = _actions[i].to;
+            (bool success, bytes memory response) = to.call{value: _actions[i].value}(
                 _actions[i].data
             );
 
-            if (!success) revert ActionFailed();
+            if(!success) {
+                // If the call failed and wasn't allowed in allowFailureMap, revert.
+                if(!hasBit(allowFailureMap, uint8(i))) {
+                    revert ActionFailed(i);
+                }
 
+                // If the call failed, but was allowed in allowFailureMap, store that 
+                // this specific action has actually failed.
+                failureMap = flipBit(failureMap, uint8(i));
+            }
+            
             execResults[i] = response;
 
             unchecked {
@@ -182,9 +220,7 @@ contract DAO is
             }
         }
 
-        emit Executed(msg.sender, callId, _actions, execResults);
-
-        return execResults;
+        emit Executed(msg.sender, callId, _actions, failureMap, execResults);
     }
 
     /// @inheritdoc IDAO
@@ -286,6 +322,28 @@ contract DAO is
         emit StandardCallbackRegistered(_interfaceId, _callbackSelector, _magicNumber);
     }
 
+    /// @inheritdoc IEIP4824
+    function daoURI() external view returns (string memory) {
+        return _daoURI;
+    }
+
+    /// @notice Updates the set DAO uri to a new value.
+    /// @param newDaoURI The new DAO uri to be set.
+    function setDaoURI(string calldata newDaoURI)
+        external
+        auth(address(this), SET_METADATA_PERMISSION_ID)
+    {
+        _setDaoURI(newDaoURI);
+    }
+
+    /// @notice Sets the new DAO uri and emits the associated event.
+    /// @param daoURI_ The new DAO uri.
+    function _setDaoURI(string calldata daoURI_) internal {
+        _daoURI = daoURI_;
+
+        emit NewURI(daoURI_);
+    }
+
     /// @notice This empty reserved space is put in place to allow future versions to add new variables without shifting down storage in the inheritance chain (see [OpenZepplins guide about storage gaps](https://docs.openzeppelin.com/contracts/4.x/upgradeable#storage_gaps)).
-    uint256[48] private __gap;
+    uint256[47] private __gap;
 }