+++ packages/contracts/test/voting/multisig.ts
@@ -18,6 +18,7 @@ import {
   getTime,
   setTimeForNextBlock,
   timestampIn,
+  toBytes32,
 } from '../test-utils/voting';
 import {shouldUpgradeCorrectly} from '../test-utils/uups-upgradeable';
 import {UPGRADE_PERMISSIONS} from '../test-utils/permissions';
@@ -347,7 +348,7 @@ describe('Multisig', function () {
       );
       await ethers.provider.send('evm_mine', []);
 
-      let proposalId0 = await multisig.callStatic.createProposal(
+      const proposalId0 = await multisig.callStatic.createProposal(
         dummyMetadata,
         dummyActions,
         false,
@@ -367,7 +368,7 @@ describe('Multisig', function () {
         )
       ).to.not.be.reverted;
 
-      let proposalId1 = await multisig.callStatic.createProposal(
+      const proposalId1 = await multisig.callStatic.createProposal(
         dummyMetadata,
         dummyActions,
         false,
@@ -389,8 +390,8 @@ describe('Multisig', function () {
         multisigSettings
       );
 
-      let startDate = await timestampIn(1000);
-      let endDate = await timestampIn(5000);
+      const startDate = await timestampIn(1000);
+      const endDate = await timestampIn(5000);
       await expect(
         multisig
           .connect(signers[0])
@@ -419,8 +420,8 @@ describe('Multisig', function () {
       });
 
       it('creates a proposal when unlisted accounts are allowed', async () => {
-        let startDate = await timestampIn(1000);
-        let endDate = await timestampIn(5000);
+        const startDate = await timestampIn(1000);
+        const endDate = await timestampIn(5000);
 
         await expect(
           multisig
@@ -481,8 +482,8 @@ describe('Multisig', function () {
       });
 
       it('creates a proposal successfully and does not approve if not specified', async () => {
-        let startDate = await timestampIn(1000);
-        let endDate = await timestampIn(5000);
+        const startDate = await timestampIn(1000);
+        const endDate = await timestampIn(5000);
 
         await ethers.provider.send('evm_setNextBlockTimestamp', [startDate]);
 
@@ -516,8 +517,7 @@ describe('Multisig', function () {
         );
         expect(proposal.parameters.startDate).to.equal(startDate);
         expect(proposal.parameters.endDate).to.equal(endDate);
-        expect(proposal.tally.approvals).to.equal(0);
-        expect(proposal.tally.addresslistLength).to.equal(1);
+        expect(proposal.approvals).to.equal(0);
         expect(proposal.actions.length).to.equal(0);
 
         expect(await multisig.canApprove(id, signers[0].address)).to.be.true;
@@ -525,8 +525,8 @@ describe('Multisig', function () {
       });
 
       it('creates a proposal successfully and approves if specified', async () => {
-        let startDate = await timestampIn(3000);
-        let endDate = await timestampIn(5000);
+        const startDate = await timestampIn(3000);
+        const endDate = await timestampIn(5000);
 
         await ethers.provider.send('evm_setNextBlockTimestamp', [startDate]);
 
@@ -555,7 +555,7 @@ describe('Multisig', function () {
         );
         expect(proposal.parameters.startDate).to.equal(startDate);
         expect(proposal.parameters.endDate).to.equal(endDate);
-        expect(proposal.tally.approvals).to.equal(1);
+        expect(proposal.approvals).to.equal(1);
       });
 
       it('increases the proposal count', async () => {
@@ -589,7 +589,7 @@ describe('Multisig', function () {
       await expect(
         multisig.createProposal(dummyMetadata, dummyActions, true, false, 5, 0)
       )
-        .to.be.revertedWithCustomError(multisig, 'InvalidStartDate')
+        .to.be.revertedWithCustomError(multisig, 'DateOutOfBounds')
         .withArgs(timeStamp, 5);
     });
 
@@ -599,7 +599,7 @@ describe('Multisig', function () {
       await expect(
         multisig.createProposal(dummyMetadata, dummyActions, true, false, 0, 5)
       )
-        .to.be.revertedWithCustomError(multisig, 'InvalidEndDate')
+        .to.be.revertedWithCustomError(multisig, 'DateOutOfBounds')
         .withArgs(timeStamp, 5);
     });
   });
@@ -694,23 +694,24 @@ describe('Multisig', function () {
       });
 
       it('reverts if minimal approval is not met yet', async () => {
-        expect(await multisig.approvals(id)).to.eq(0);
+        const proposal = await multisig.getProposal(id)
+        expect(proposal.approvals).to.eq(0);
         await expect(multisig.execute(id))
           .to.be.revertedWithCustomError(multisig, 'ProposalExecutionForbidden')
           .withArgs(id);
       });
 
       it('approves with the msg.sender address', async () => {
-        expect((await multisig.getProposal(id)).tally.approvals).to.equal(0);
+        expect((await multisig.getProposal(id)).approvals).to.equal(0);
 
-        let tx = await multisig.connect(signers[0]).approve(id, false);
+        const tx = await multisig.connect(signers[0]).approve(id, false);
 
         const event = await findEvent(tx, 'Approved');
         expect(event.args.proposalId).to.eq(id);
         expect(event.args.approver).to.not.eq(multisig.address);
         expect(event.args.approver).to.eq(signers[0].address);
 
-        expect((await multisig.getProposal(id)).tally.approvals).to.equal(1);
+        expect((await multisig.getProposal(id)).approvals).to.equal(1);
       });
 
       it("reverts if the proposal hasn't started yet", async () => {
@@ -758,9 +759,8 @@ describe('Multisig', function () {
 
     describe('canExecute:', async () => {
       it('returns `false` if the proposal has not reached the minimum approval yet', async () => {
-        expect(await multisig.approvals(id)).to.be.lt(
-          (await multisig.getProposal(id)).parameters.minApprovals
-        );
+        const proposal = await multisig.getProposal(id);
+        expect(proposal.approvals).to.be.lt(proposal.parameters.minApprovals);
 
         expect(await multisig.canExecute(id)).to.be.false;
       });
@@ -841,9 +841,7 @@ describe('Multisig', function () {
         expect(proposal.parameters.minApprovals).to.equal(
           multisigSettings.minApprovals
         );
-        expect(await multisig.approvals(id)).to.be.eq(
-          multisigSettings.minApprovals
-        );
+        expect(proposal.approvals).to.be.eq(multisigSettings.minApprovals);
 
         expect(await multisig.canExecute(id)).to.be.true;
         await expect(multisig.execute(id)).to.not.be.reverted;
@@ -857,9 +855,7 @@ describe('Multisig', function () {
         expect(proposal.parameters.minApprovals).to.equal(
           multisigSettings.minApprovals
         );
-        expect(await multisig.approvals(id)).to.be.eq(
-          multisigSettings.minApprovals
-        );
+        expect(proposal.approvals).to.be.eq(multisigSettings.minApprovals);
 
         expect(await multisig.canExecute(id)).to.be.true;
         expect(await multisig.isListed(signers[9].address)).to.be.false; // signers[9] is not listed
@@ -888,7 +884,7 @@ describe('Multisig', function () {
           const event = await findEvent(tx, DAO_EVENTS.EXECUTED);
 
           expect(event.args.actor).to.equal(multisig.address);
-          expect(event.args.callId).to.equal(id);
+          expect(event.args.callId).to.equal(toBytes32(id));
           expect(event.args.actions.length).to.equal(1);
           expect(event.args.actions[0].to).to.equal(dummyActions[0].to);
           expect(event.args.actions[0].value).to.equal(dummyActions[0].value);